parser.y: fix yyscan_t not defined with bison 3.0
The code worked great with bison 2.5 but something changed between
that release and 3.0, and yyscan_t was no longer defined. I couldn't
find much on the topic since re-entrant parsers in C aren't too
common. The only report of a similar problem was with envytools.
Searching through their git repo I found that they solved the problem
by using a void*. It seems to work, so I'll go with it for now.