Makefile.am: fix `make distcheck`
The testsuite directory wasn't being created in the build tree
which caused the output file not to be created causing all the
tests to fail. The fix was to move the test files into the top
level directory. If it grows too big, in the future I may split
things out again, but with autoconf/automake, it seems to be
more trouble than it is worth.